Removes duplicate post method from mplane.client
There really shouldn't be two functions with the same name. The second
function in the file was removed as the first one has seen a greater
number and more recent edits than the second one. The old second
function will still be in the git history if we need it.