Adds a default _ssn to BaseClient that could be useful for unit tests
BaseClient has functions that call _ssn and expect it to exist. In order
to perform unit testing on BaseClient, these calls would need to not
fail. By setting the default serial number to -1, using this as a known
bad value, this can be checked to ensure that it is not still -1 when
it's actually being used.